+++ src/main/java/fr/bankwiz/server/model/User.java
@@ -15,7 +15,7 @@ public class User {
     @Id
     @Column(name = "USER_ID", nullable = false, updatable = false, insertable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Long userId;
+    protected Integer userId;
 
     @Column(name = "FIRST_NAME", nullable = false, length = 60)
     private String firstName;
